OO-106: moved buchhaltung repositories into package
