+++ src/main/scala/ch/openolitor/buchhaltung/BuchhaltungDeleteService.scala
@@ -35,6 +35,8 @@ import ch.openolitor.core.domain.EntityStore._
 import ch.openolitor.buchhaltung.models._
 import scala.concurrent.ExecutionContext.Implicits.global
 import ch.openolitor.core.models.PersonId
+import ch.openolitor.buchhaltung.repositories.DefaultBuchhaltungWriteRepositoryComponent
+import ch.openolitor.buchhaltung.repositories.BuchhaltungWriteRepositoryComponent
 
 object BuchhaltungDeleteService {
   def apply(implicit sysConfig: SystemConfig, system: ActorSystem): BuchhaltungDeleteService = new DefaultBuchhaltungDeleteService(sysConfig, system)
