+++ src/main/scala/ch/openolitor/buchhaltung/repositories/BuchhaltungWriteRepositoryComponent.scala
@@ -20,14 +20,18 @@
 * with this program. If not, see http://www.gnu.org/licenses/                 *
 *                                                                             *
 \*                                                                           */
-package ch.openolitor.buchhaltung
+package ch.openolitor.buchhaltung.repositories
 
 import akka.actor.ActorSystem
+import ch.openolitor.core.DefaultActorSystemReference
+import ch.openolitor.core.AkkaEventStream
 
-trait BuchhaltungReadRepositoryComponent {
-  val buchhaltungReadRepository: BuchhaltungReadRepository
+trait BuchhaltungWriteRepositoryComponent {
+  val buchhaltungWriteRepository: BuchhaltungWriteRepository
 }
 
-trait DefaultBuchhaltungReadRepositoryComponent extends BuchhaltungReadRepositoryComponent {
-  override val buchhaltungReadRepository: BuchhaltungReadRepository = new BuchhaltungReadRepositoryImpl
+trait DefaultBuchhaltungWriteRepositoryComponent extends BuchhaltungWriteRepositoryComponent {
+  val system: ActorSystem
+
+  override val buchhaltungWriteRepository: BuchhaltungWriteRepository = new DefaultActorSystemReference(system) with BuchhaltungWriteRepositoryImpl with AkkaEventStream
 }
